More logging, more traps for null cache, no solution.
The issue in @m.rae's logs is that an Outlook item that was present at startup and which had the correct duration when we first saw it had it's duration set to zero while Outlook was resolving its recipient references. At that point we don't have a cache of the item because we haven't yet done anything to the item, so we cannot detect and fix duration set to zero.
So my proposed solution is that we should cache the values of the item the first moment we see it, and test for duration-set-to-zero each time we save it.
4:01 PM
I don't know whether this is a good solution. It's adding complexity and it doesn't feel to me that we've found the underlying cause. But it's the best I can come up with now.